Added Temp/Perm Vehicle Spawn selection with images
Need to add this line in the description.ext, at the bottom of the file.
#include "admintools\dialog.hpp"

Selecting the menu the 1st time will take some time to load, but after
that it will pop out fast.
There are 3 buttons on the dialog window, "Temp", "Perm", "Close". Temp
spawn will make a temporary vehicle...etc