Fix wait until issue
Removed wait until from the readme and set adminlist to be started at
the top of the init using preprocess file so that the adminlist is done
before the player is even loaded. Makes the install a little longer, but
it is the best way to do it.

Also updated readme for better update method.

Fixes #177